feat(suite): Add IconButton to Warning
