+++ packages/components/src/components/Warning/Warning.tsx
@@ -1,36 +1,28 @@
 import { ReactNode } from 'react';
-import styled, { DefaultTheme, css, useTheme } from 'styled-components';
+import styled, { css, useTheme } from 'styled-components';
 
 import { Icon, IconType } from '../assets/Icon/Icon';
 import { variables } from '../../config';
-import {
-    CSSColor,
-    Color,
-    Elevation,
-    borders,
-    mapElevationToBackgroundToken,
-    spacingsPx,
-    typography,
-    spacings,
-} from '@trezor/theme';
-import { UIVariant } from '../../config/types';
-import { Button, ButtonProps, Row, TransientProps, useElevation } from '../..';
+import { Elevation, borders, spacingsPx, typography, spacings } from '@trezor/theme';
+import { Row, TransientProps, useElevation } from '../..';
 import { FrameProps, FramePropsKeys, withFrameProps } from '../common/frameProps';
-import { useWarningContextContext, WarningContext } from './WarningContext';
+import { WarningContext } from './WarningContext';
+import { WarningButton } from './WarningButton';
+import { WarningVariant } from './types';
+import { DEFAULT_VARIANT } from './consts';
+import { WarningIconButton } from './WarningIconButton';
+import {
+    mapVariantToBackgroundColor,
+    mapVariantToIcon,
+    mapVariantToIconColor,
+    mapVariantToTextColor,
+} from './utils';
 
 export const allowedWarningFrameProps: FramePropsKeys[] = ['margin'];
 type AllowedFrameProps = Pick<FrameProps, (typeof allowedWarningFrameProps)[number]>;
 
-export type WarningVariant = Extract<
-    UIVariant,
-    'primary' | 'secondary' | 'info' | 'warning' | 'destructive' | 'tertiary'
->;
-
-const DEFAULT_VARIANT = 'warning' as WarningVariant;
-
 export type WarningProps = AllowedFrameProps & {
     children: ReactNode;
-    extra?: ReactNode;
     className?: string;
     variant?: WarningVariant;
     rightContent?: ReactNode;
@@ -38,63 +30,6 @@ export type WarningProps = AllowedFrameProps & {
     icon?: IconType;
     filled?: boolean;
     dataTest?: string;
-};
-
-type MapArgs = {
-    $variant: WarningVariant;
-    theme: DefaultTheme;
-    $elevation: Elevation;
-};
-
-const mapVariantToBackgroundColor = ({ $variant, theme, $elevation }: MapArgs): CSSColor => {
-    const colorMap: Record<WarningVariant, Color> = {
-        primary: 'backgroundPrimarySubtleOnElevation0',
-        secondary: 'backgroundNeutralBold',
-        info: 'backgroundAlertBlueSubtleOnElevation0',
-        warning: 'backgroundAlertYellowSubtleOnElevation0',
-        destructive: 'backgroundAlertRedSubtleOnElevation0',
-        tertiary: mapElevationToBackgroundToken({ $elevation }),
-    };
-
-    return theme[colorMap[$variant]];
-};
-
-const mapVariantToTextColor = ({ $variant, theme }: MapArgs): CSSColor => {
-    const colorMap: Record<WarningVariant, Color> = {
-        primary: 'textPrimaryDefault',
-        secondary: 'textDefaultInverted',
-        info: 'textAlertBlue',
-        warning: 'textAlertYellow',
-        destructive: 'textAlertRed',
-        tertiary: 'textSubdued',
-    };
-
-    return theme[colorMap[$variant]];
-};
-const mapVariantToIconColor = ({ $variant, theme }: MapArgs): CSSColor => {
-    const colorMap: Record<WarningVariant, Color> = {
-        primary: 'iconPrimaryDefault',
-        secondary: 'iconDefaultInverted',
-        info: 'iconAlertBlue',
-        warning: 'iconAlertYellow',
-        destructive: 'iconAlertRed',
-        tertiary: 'iconSubdued',
-    };
-
-    return theme[colorMap[$variant]];
-};
-
-const mapVariantToIcon = ({ $variant }: Pick<MapArgs, '$variant'>): IconType => {
-    const iconMap: Record<WarningVariant, IconType> = {
-        primary: 'LIGHTBULB',
-        secondary: 'INFO',
-        info: 'INFO',
-        warning: 'WARNING',
-        destructive: 'WARNING',
-        tertiary: 'INFO',
-    };
-
-    return iconMap[$variant];
 };
 
 type WrapperParams = TransientProps<AllowedFrameProps> & {
@@ -165,6 +100,7 @@ export const Warning = ({
                     })}
                 />
             )}
+
             <Row justifyContent="space-between" gap={spacings.lg} flex={1}>
                 {children}
                 {rightContent && (
@@ -177,17 +113,5 @@ export const Warning = ({
     );
 };
 
-const WarningButton = ({ children, ...rest }: ButtonProps) => {
-    const { variant } = useWarningContextContext();
-    const value = { variant: DEFAULT_VARIANT };
-
-    return (
-        <WarningContext.Provider value={value}>
-            <Button {...rest} variant={rest.variant ?? variant} size={rest.size ?? 'small'}>
-                {children}
-            </Button>
-        </WarningContext.Provider>
-    );
-};
-
 Warning.Button = WarningButton;
+Warning.IconButton = WarningIconButton;
