+++ packages/components/src/components/Warning/consts.ts
@@ -1,0 +1,3 @@
+import { WarningVariant } from './types';
+
+export const DEFAULT_VARIANT = 'warning' as WarningVariant;
