+++ packages/components/src/components/Warning/types.ts
@@ -1,0 +1,6 @@
+import { UIVariant } from '../../config/types';
+
+export type WarningVariant = Extract<
+    UIVariant,
+    'primary' | 'secondary' | 'info' | 'warning' | 'destructive' | 'tertiary'
+>;
